un comment failed test (#69)
* fix: un-comment failed test

* file: place test into specific package

* test integration: fix assert exclude pair logic

Junit isNotIn cannot handle array
