+++ src/test/java/org/fit/ssapp/gt/GTPayoffUnitTest.java
@@ -1,4 +1,4 @@
-package org.fit.ssapp.service;
+package org.fit.ssapp.gt;
 
 import static org.junit.jupiter.api.Assertions.*;
