+++ src/test/java/org/fit/ssapp/service/st/OTMStableProblemServiceTest.java
@@ -1,3 +1,4 @@
+package org.fit.ssapp.service.st;
 //package com.example.SS2_Backend.service;
 //
 //import com.example.SS2_Backend.dto.request.StableMatchingProblemDto;
